Fix database schema and API endpoints - salary column DOUBLE type, fix security config, add health controller

+++ backend/src/main/resources/db/migration/V1__Initial_Schema.sql
@@ -19,7 +19,7 @@ CREATE TABLE employees (
     employee_id VARCHAR(50) NOT NULL UNIQUE,
     department VARCHAR(100),
     position VARCHAR(100),
-    salary DECIMAL(10, 2),
+    salary DOUBLE,
     hire_date DATE,
     phone VARCHAR(20),
     address VARCHAR(255),
